An u_int8_t is enough as prefixlen. Another lint warning gone.

+++ usr.sbin/ripd/kroute.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: kroute.c,v 1.10 2007/10/17 20:18:11 claudio Exp $ */
+/*	$OpenBSD: kroute.c,v 1.11 2007/10/18 17:00:59 deraadt Exp $ */
 
 /*
  * Copyright (c) 2004 Esben Norby <norby@openbsd.org>
@@ -562,7 +562,7 @@ kif_validate(int ifindex)
 struct kroute_node *
 kroute_match(in_addr_t key)
 {
-	int			 i;
+	u_int8_t		 i;
 	struct kroute_node	*kr;
 
 	/* we will never match the default route */
